+++ routerun-app/src/section/Goal.jsx
@@ -1,0 +1,3 @@
+export const Goal = () => {
+  return <></>;
+};
