+++ routerun-app/src/section/Start.jsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 
-export const FirstViewDemo = ({ setLocation }) => {
+export const Start = ({ setLocation }) => {
   useEffect(() => {
     if (!navigator.geolocation) {
       console.error("Geolocation はこのブラウザでサポートされていません。");
@@ -11,9 +11,7 @@ export const FirstViewDemo = ({ setLocation }) => {
       (position) => {
         const lat = position.coords.latitude;
         const lng = position.coords.longitude;
-        console.log("現在地:", { lat, lng });
-        //勝手に追加しました
-        setLocation({ lat, lng }); // 取得した座標を setLocation で更新
+        setLocation({ lat, lng });
       },
       (error) => {
         console.error("現在地の取得に失敗:", error);
